+++ code/cmt/python/src/trainableinterface.cpp
@@ -8,6 +8,10 @@ using Eigen::Dynamic;
 using Eigen::Matrix;
 using Eigen::Map;
 
+#include <iostream>
+using std::cout;
+using std::endl;
+
 Trainable::Parameters* PyObject_ToParameters(PyObject* parameters) {
 	return PyObject_ToParameters(parameters, new Trainable::Parameters);
 }
@@ -46,7 +50,7 @@ Trainable::Parameters* PyObject_ToParameters(
 			if(PyFloat_Check(threshold))
 				params->threshold = PyFloat_AsDouble(threshold);
 			else if(PyInt_Check(threshold))
-				params->threshold = static_cast<double>(PyFloat_AsDouble(threshold));
+				params->threshold = static_cast<double>(PyInt_AsLong(threshold));
 			else
 				throw Exception("threshold should be of type `float`.");
 
@@ -192,6 +196,14 @@ PyObject* Trainable_train(
 		input_val = 0;
 	if(output_val == Py_None)
 		output_val = 0;
+
+	if((input_val && PyDict_Check(input_val)) || (output_val && PyDict_Check(output_val))) {
+		// for some reason PyArray_FROM_OTF segfaults when input_val is a dictionary
+		Py_DECREF(input);
+		Py_DECREF(output);
+		PyErr_SetString(PyExc_TypeError, "Validation data has to be stored in NumPy arrays.");
+		return 0;
+	}
 
 	if(input_val || output_val) {
 		input_val = PyArray_FROM_OTF(input_val, NPY_DOUBLE, NPY_F_CONTIGUOUS | NPY_ALIGNED);
@@ -580,6 +592,75 @@ PyObject* Trainable_parameter_gradient(
 		Py_DECREF(input);
 		Py_DECREF(output);
 		Py_XDECREF(x);
+		PyErr_SetString(PyExc_RuntimeError, exception.message());
+		return 0;
+	}
+}
+
+
+
+const char* Trainable_fisher_information_doc =
+	"_parameter_gradient(self, input, output, x=None, parameters=None)\n"
+	"\n"
+	"Estimates the Fisher information matrix of the parameters as returned by L{_parameters()}.\n"
+	"\n"
+	"@type  input: C{ndarray}\n"
+	"@param input: inputs stored in columns\n"
+	"\n"
+	"@type  output: C{ndarray}\n"
+	"@param output: outputs stored in columns\n"
+	"\n"
+	"@type  parameters: C{dict}\n"
+	"@param parameters: a dictionary containing hyperparameters\n"
+	"\n"
+	"@rtype: C{ndarray}\n"
+	"@return: the Fisher information matrix";
+
+PyObject* Trainable_fisher_information(
+	TrainableObject* self,
+	PyObject* args,
+	PyObject* kwds,
+	Trainable::Parameters* (*PyObject_ToParameters)(PyObject*))
+{
+	const char* kwlist[] = {"input", "output", "parameters", 0};
+
+	PyObject* input;
+	PyObject* output;
+	PyObject* parameters = 0;
+
+	// read arguments
+	if(!PyArg_ParseTupleAndKeywords(args, kwds, "OO|O", const_cast<char**>(kwlist),
+		&input, &output, &parameters))
+		return 0;
+
+	// make sure data is stored in NumPy array
+	input = PyArray_FROM_OTF(input, NPY_DOUBLE, NPY_F_CONTIGUOUS | NPY_ALIGNED);
+	output = PyArray_FROM_OTF(output, NPY_DOUBLE, NPY_F_CONTIGUOUS | NPY_ALIGNED);
+
+	if(!input || !output) {
+		Py_XDECREF(input);
+		Py_XDECREF(output);
+		PyErr_SetString(PyExc_TypeError, "Data has to be stored in NumPy arrays.");
+		return 0;
+	}
+
+	try {
+		Trainable::Parameters* params = PyObject_ToParameters(parameters);
+
+		MatrixXd fisherInformation = self->distribution->fisherInformation(
+			PyArray_ToMatrixXd(input),
+			PyArray_ToMatrixXd(output),
+			*params);
+
+		delete params;
+
+		Py_DECREF(input);
+		Py_DECREF(output);
+
+		return PyArray_FromMatrixXd(fisherInformation);
+	} catch(Exception exception) {
+		Py_DECREF(input);
+		Py_DECREF(output);
 		PyErr_SetString(PyExc_RuntimeError, exception.message());
 		return 0;
 	}